refactor(scripts): organize utilities
