+++ package.json
@@ -30,6 +30,9 @@
     "type-fest": "^4.30.2",
     "usercss-meta": "0.12.0"
   },
+  "devDependencies": {
+    "@types/less": "^3.0.8"
+  },
   "engines": {
     "node": ">=23.6.0"
   },
