+++ scripts/utils/data.ts
@@ -1,10 +1,10 @@
-import type { CategoriesSchema, UserstylesSchema } from "./types/mod.ts";
+import type { CategoriesSchema, UserstylesSchema } from "../types/mod.ts";
 import type { SetRequired } from "type-fest/source/set-required.d.ts";
 import {
   CATEGORIES_SCHEMA,
   REPO_ROOT,
   USERSTYLES_SCHEMA,
-} from "./constants.ts";
+} from "../constants.ts";
 
 import * as yaml from "@std/yaml";
 import path from "node:path";
@@ -12,8 +12,12 @@ import path from "node:path";
 import Ajv, { type Schema } from "ajv";
 import { log } from "./logger.ts";
 import { sprintf } from "@std/fmt/printf";
-import { Octokit } from "@octokit/rest";
-import { readDirSync, readTextFileSync } from "./utils/fs.ts";
+import { readDirSync, readTextFileSync } from "./fs.ts";
+
+type Userstyles = SetRequired<
+UserstylesSchema.UserstylesSchema,
+"userstyles" | "collaborators"
+>;
 
 /**
  * @param content A string of YAML content
@@ -56,7 +60,7 @@ export function validateYaml<T>(
 
 /**
  * Utility function that calls {@link validateYaml} on the userstyles.yml file.
- * Fails when data.userstyles is undefined.
+ * Requires the `userstyles` and `collaborators` fields.
  */
 export function getUserstylesData(): Userstyles {
   const content = readTextFileSync(
@@ -102,8 +106,7 @@ export function getUserstylesData(): Userstyles {
 }
 
 /**
- * Utility function that calls {@link validateYaml} on the ports.yml file.
- * Fails when data.userstyles is undefined.
+ * Utility function that calls {@link validateYaml} on the categories.yml file.
  */
 export async function getCategoriesData(): Promise<
   CategoriesSchema.CategoryDefinitions
@@ -121,36 +124,6 @@ export async function getCategoriesData(): Promise<
   return data;
 }
 
-/**
- * Utility function that formats a list of items into the "x, y, ..., and z" format.
- * @example
- * formatListOfItems(['x']); // 'x'
- * @example
- * formatListOfItems(['x', 'y']); // 'x and y'
- * @example
- * formatListOfItems(['x', 'y', 'z']); // 'x, y, and z'
- */
-export function formatListOfItems(items: unknown[]): string {
-  // If there are no items, return an empty string.
-  if (items.length === 0) return "";
-  // If there are two items, connect them with an "and".
-  if (items.length === 2) return items.join(" and ");
-  // Otherwise, there is either just one item or more than two items.
-  return items.reduce((prev, curr, idx, arr) => {
-    // If this is the first item of the items we are looping through, set our initial string to it.
-    if (idx === 0) return curr;
-    // If this is the last one, add a comma (Oxford commas are amazing) followed by "and" and the item to the string.
-    if (curr === arr.at(-1)) return prev + `, and ${curr}`;
-    // Otherwise, it is some item in the middle of the list and we can just add it as a comma followed by the item to the string.
-    return prev + `, ${curr}`;
-  }) as string;
-}
-
-type Userstyles = SetRequired<
-  UserstylesSchema.UserstylesSchema,
-  "userstyles" | "collaborators"
->;
-
 export function getUserstylesFiles(): string[] {
   const files: string[] = [];
   for (const dir of readDirSync(path.join(REPO_ROOT, "styles"))) {
@@ -160,46 +133,4 @@ export function getUserstylesFiles(): string[] {
     );
   }
   return files;
-}
-
-export function getAuthenticatedOctokit() {
-  return new Octokit({ auth: process.env["GITHUB_TOKEN"] });
-}
-
-export type UserstylesTeam = "userstyles-staff" | "userstyles-maintainers";
-
-export async function getUserstylesTeamMembers(
-  octokit: Octokit,
-  team: UserstylesTeam,
-): Promise<string[]> {
-  const members = await octokit.teams.listMembersInOrg({
-    org: "catppuccin",
-    team_slug: team,
-    per_page: 100,
-  });
-  return members.data.map(({ login }) => login.toLowerCase());
-}
-
-export async function addUserstylesTeamMember(
-  octokit: Octokit,
-  team: UserstylesTeam,
-  username: string,
-) {
-  await octokit.teams.addOrUpdateMembershipForUserInOrg({
-    org: "catppuccin",
-    team_slug: team,
-    username,
-  });
-}
-
-export async function removeUserstylesTeamMember(
-  octokit: Octokit,
-  team: UserstylesTeam,
-  username: string,
-) {
-  await octokit.teams.removeMembershipForUserInOrg({
-    org: "catppuccin",
-    team_slug: team,
-    username,
-  });
 }
